Added BukkitInterface.getOnlinePlayers()
The problem with using Bukkit.getOnlinePlayers() is that the API has two
different return values depending on which version of the API that you're
compiling against. Then, at runtime, the server may have a completely
different, unsupport version.

BukkitInterface is the solution to interfacing with incompatible versions
of the Bukkit API.

Use Bukkit.getOnlinePlayers() and your plugin can potentially fail at
runtime.

Use BukkitInterface.getOnlinePlayers() to avoid this problem.

These are the possible solutions:
- use reflection
- return an rx.Observable<Player>
- use an interface with 2 different implementations

This resource uses the last approach: It defines an interface
OnlinePlayers with 2 different implementations: OnlinePlayersArray &
OnlinePlayersCollections.

OlinePlayersCollections is used for 1.7.10+
OlinePlayersArray is used for 1.7.9-

BukkitInterface detects the runtime environment (what version of Spigot or
Craftbukkit the server is running) and statically instantiates one of the
implementations. BukkitInterface then uses this to power getOlinePlayers()

Admittedly, using reflection would have been easier. And returning an
"event stream" or "data pipeline" via an rx.Observable would have been an
interesting solution (and possibly allow for safe, asynchronous operations).

I will add other implementations to the BukkitInterface as the need
arises.