BukkitInterface v2.0
- Now uses mc.euro:Version v2.0

The major change here is that VersionFactory.getNmsVersion() no longer
returns "pre" or "vpre" for old Minecraft versions. This responsibility is
now up to the end-user.

VersionFactory.getNmsVersion() :

    package net.minecraft.server.{version}.*;

The {version} format is vX_Y_Z. If this {version} package does not exist,
then it is constructed. This is done to protect against possible changes
where Spigot no longer uses this versioned package structure.

So, this doesn't effect the BukkitInterface at all. But it does effect any
projects that uses BukkitInterface's Version & VersionFactory classes:

Like, BattleBukkitLib.