BukkitInterface v2.0.1
- now uses mc.euro:Version v2.0.1

This fixes an edge case in version.compareTo() between (for example):

v2.0.0.1
v2.0

Normally, for a worst case scenario, v2.0 would get a bump to v2.0.1, in
which case, compareTo() would succeed. However, for the example of v2.0
verse v2.0.0.1, they would get converted into two arrays:

[2][0][0][1]
[2][0]

And the shorter array would trigger an IndexOutOfBoundsException and get
handled incorrectly for this edge case (ignore the fourth digit, and
simply return the difference between the third digits: Zero -- which
incorrectly labels them the same versions).

This update employs a much cleaner solution. And the edge case has been
added to the unit tests.